SOCP collocation in different shard
